+++ changelog.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 4.1.3
+
+### Fixed
+
+- Block highlight lines now render correctly when Iris shaders are enabled.
+
+## 3.1.4
+
+### Fixed
+
+- Block highlight lines now render correctly when Iris shaders are enabled.
+
 ## 4.1.2
 
 ### Fixed